backend/requirements.txt
fastapi
uvicorn
numpy
pandas
requests
python-multipart
torch
sentence-transformers
Pillow
ultralytics
opencv-python
